sec manager, move pairing_status.hpp
