+++ bluetoe/link_state.hpp
@@ -1,7 +1,7 @@
 #ifndef BLUETOE_LINK_STATE_HPP
 #define BLUETOE_LINK_STATE_HPP
 
-#include <bluetoe/sm/pairing_status.hpp>
+#include <bluetoe/pairing_status.hpp>
 #include <bluetoe/codes.hpp>
 
 namespace bluetoe {
